requirements.txt
beautifulsoup4>=4.10.0
requests>=2.25.0
selenium>=4.0.0
typer>=0.4.0
PyYAML>=6.0
Flask>=2.0.0
pandas>=1.3.0
jsonschema>=4.0.0
lxml>=4.6.0  # Optional: Needed if you implement XPath in html_scraper
